fix: matching socket to nvim parent only
There was an unintentional check to see if the process id of the neovim socket
(--embed) is the "client" neovim process, which shouldn't ever be the case (I
think). The behavior was a false positive that opens up the file from any pane
in a running instance of neovim in a different tmux window. I went ahead and
added this functionality, since I think it's useful. Now, it ton will try to
open in a nvim instance in the current window, or fallback to the first nvim
instance it finds.