feat: configure prompting with nvims
Add a config to prioritize the current window if there are multiple instances of
nvim running in the session (in other windows). If not prioritized then user
will be prompted with a menu to select an instance.

Note, if more than one nvim is in the same window and prioritize is true, then
it will automatically choose the first instance found.